requirements.txt
Django==4.1.7
python-decouple==3.8
requests==2.28.2

lyricsgenius==3.0.1
python3-discogs-client==2.6